Move `Discovery` class into `common` folder
We'll also use this for discovering query files for the new Queries panel
